requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
canvasapi==2.1.0
certifi==2020.12.5
chardet==4.0.0
commonmark==0.9.1
dictdiffer==0.8.1
idna==2.10
notion==0.0.27
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
soupsieve==2.1
text-unidecode==1.3
todoist-python==8.1.3
tzlocal==2.1
urllib3==1.26.2
